Allow feedback to be sent even if user is not authenticated
